Fix SQLDelight WASM schema not created on startup
WebWorkerDriver was created but BillibDatabase.Schema.create() was never
called, so tables didn't exist. Fix by awaiting async schema creation
before Koin initializes.

- initializeDatabaseDriver() suspend fn creates driver and awaits schema
- main() uses MainScope().launch to run async init before startKoin
- NoOpSqlDriver/NoOpCursor moved to commonMain as internal classes for
  optional fallback use; empty bodies changed to = Unit for detekt
